+++ plugin-server/tests/postgres/vm.extra-lazy.test.ts
@@ -1,0 +1,99 @@
+import * as fetch from 'node-fetch'
+
+import { createHub } from '../../src/utils/db/hub'
+import { pluginDigest } from '../../src/utils/utils'
+import { LazyPluginVM } from '../../src/worker/vm/lazy'
+import { plugin60, pluginConfig39 } from '../helpers/plugins'
+import { resetTestDatabase } from '../helpers/sql'
+import { Hub } from './../../src/types'
+
+describe('VMs are extra lazy 💤', () => {
+    let hub: Hub
+    let closeHub: () => Promise<void>
+
+    beforeEach(async () => {
+        ;[hub, closeHub] = await createHub()
+    })
+
+    afterEach(async () => {
+        await closeHub()
+        jest.clearAllMocks()
+    })
+    test('VM with scheduled tasks gets setup immediately', async () => {
+        const indexJs = `
+        export async function runEveryMinute () {
+            console.log('haha')
+        }
+
+        export async function setupPlugin () {
+            await fetch('https://onevent.com/')
+        }
+    `
+        await resetTestDatabase(indexJs)
+        const lazyVm = new LazyPluginVM()
+        jest.spyOn(lazyVm, 'setupPluginIfNeeded')
+
+        const pluginConfig = { ...pluginConfig39, vm: lazyVm }
+        await lazyVm.initialize!(hub, pluginConfig, indexJs, pluginDigest(plugin60))
+
+        expect(lazyVm.ready).toEqual(true)
+        expect(lazyVm.setupPluginIfNeeded).not.toHaveBeenCalled()
+        expect(fetch).toHaveBeenCalledWith('https://onevent.com/')
+    })
+
+    test('VM with jobs gets setup immediately', async () => {
+        const indexJs = `
+        export async function setupPlugin () {
+            await fetch('https://onevent.com/')
+        }
+
+        export const jobs = {
+            test: (payload, meta) => {
+                console.log(payload)
+            }
+        }
+    `
+        await resetTestDatabase(indexJs)
+        const lazyVm = new LazyPluginVM()
+        jest.spyOn(lazyVm, 'setupPluginIfNeeded')
+
+        const pluginConfig = { ...pluginConfig39, vm: lazyVm }
+        await lazyVm.initialize!(hub, pluginConfig, indexJs, pluginDigest(plugin60))
+
+        expect(lazyVm.ready).toEqual(true)
+        expect(lazyVm.setupPluginIfNeeded).not.toHaveBeenCalled()
+        expect(fetch).toHaveBeenCalledWith('https://onevent.com/')
+    })
+
+    test('VM without tasks delays setup until necessary', async () => {
+        const indexJs = `
+        export async function setupPlugin () {
+            await fetch('https://onevent.com/')
+        }
+
+        export async function onEvent () {
+            
+        }
+    `
+        await resetTestDatabase(indexJs)
+        const lazyVm = new LazyPluginVM()
+        jest.spyOn(lazyVm, 'setupPluginIfNeeded')
+
+        const pluginConfig = { ...pluginConfig39, vm: lazyVm, plugin: plugin60 }
+        await lazyVm.initialize!(hub, pluginConfig, indexJs, pluginDigest(plugin60))
+
+        expect(lazyVm.ready).toEqual(false)
+        expect(lazyVm.setupPluginIfNeeded).not.toHaveBeenCalled()
+        expect(fetch).not.toHaveBeenCalled()
+
+        await lazyVm.getOnAction()
+        expect(lazyVm.ready).toEqual(false)
+        expect(lazyVm.setupPluginIfNeeded).not.toHaveBeenCalled()
+        expect(fetch).not.toHaveBeenCalled()
+
+        await lazyVm.getOnEvent()
+        expect(lazyVm.ready).toEqual(true)
+        expect(lazyVm.setupPluginIfNeeded).toHaveBeenCalled()
+        expect(fetch).toHaveBeenCalledWith('https://onevent.com/')
+    })
+})
